+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Fix bug where secrets were not attached to functions when using older Functions SDK (#4797).
 - Increase timeout of the Functions Emulator to wait for spawned process to initialize (#4944).
+- Fix bug where Storage Emulator did not convert non-string metadata value as key (#4955).